requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.109.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.27.1
